refactor(utils): reorganize constants and mappings into dedicated modules
Moved constants and mappings from `utils/constants.py` to separate modules under `utils/mappings` for better maintainability. Updated imports across the codebase to reflect these changes. This restructuring improves code organization and makes it easier to manage and extend mappings in the future.
