+++ belso/utils/__init__.py
@@ -1,7 +1,7 @@
 # belso.utils.__init__
 
 from belso.utils.logging import get_logger
-from belso.utils.constants import FORMATS
+from belso.utils.formats import FORMATS
 from belso.utils.detecting import detect_schema_format
 
 __all__ = [
